Backend/requirements.txt
# For Backend dependencies
# 
# IMPORTANT: For GPU support, first install PyTorch with CUDA:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
# Then install the rest: pip install -r requirements.txt
#
fastapi[all]
uvicorn[standard]
sqlalchemy
alembic
websockets
asyncio
python-dotenv
python-multipart
bcrypt
cloudinary
fastapi-mail
databases
slowapi
asyncpg
psycopg2-binary==2.9.11
celery
redis
python-jose[cryptography]
nest-asyncio
pinecone
# PyTorch with CUDA support - install before sentence-transformers
# Install with: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
torch>=2.5.1
torchvision>=0.20.1
torchaudio>=2.5.1
sentence-transformers
psutil
kombu
google-genai
apscheduler
asgiref
httpx
openai
exponent-server-sdk